Document and restructure the build system
Writing an extensive documentation for the build system, I found that things
might become easier to descibe if I were to split up the command line
interpretation and the overall execution logic.
