+++ .travis.yml
@@ -1,6 +1,6 @@
 language: node_js
 node_js: node
-script: make alltests
+script: node make alltests
 sudo: false
 addons:
   apt:
